Don't clobber null values if array_key_exists(PHP_INT_MAX, $array)
Handle the (extremely unusual) case where the PHP_INT_MAX or
PHP_INT_MAX-1 index of the array is in use, but with a null value:
don't clobber the key in that case.  Similarly, don't clobber null
values when selecting a random place to write the array marker and
key.